+++ src/engine/components/numberGenerator/index.stories.jsx
@@ -27,8 +27,6 @@ const MARGINS = {
 };
 const one = numberGenerator(ONE, POSITION, POSITION, SIZES.one, MARGINS.one).view;
 const ten = numberGenerator(TEN, POSITION, POSITION, SIZES.ten, MARGINS.ten).view;
-const negativeOne = numberGenerator(-ONE, POSITION, POSITION, SIZES.negativeOne, MARGINS.negativeOne).view;
-const negativeTen = numberGenerator(-TEN, POSITION, POSITION, SIZES.negativeTen, MARGINS.negativeTen).view;
 
 storiesOf('engine.components.Number', module)
   .add('one digit', () => (
@@ -47,19 +45,31 @@ storiesOf('engine.components.Number', module)
       width={SIZES.ten + MARGINS.ten + MARGINS.ten}
     />
   ))
-  .add('negative one digit', () => (
-    <PixiWrapper
-      component={negativeOne}
-      isContainer={false}
-      height={SIZES.negativeOne + MARGINS.negativeOne + MARGINS.negativeOne}
-      width={SIZES.negativeOne + MARGINS.negativeOne + MARGINS.negativeOne}
-    />
-  ))
-  .add('negative two digits', () => (
-    <PixiWrapper
-      component={negativeTen}
-      isContainer={false}
-      height={SIZES.negativeTen + MARGINS.negativeTen + MARGINS.negativeTen}
-      width={SIZES.negativeTen + MARGINS.negativeTen + MARGINS.negativeTen}
-    />
-  ));
+  .add('happy negative one digit', () => {
+    const negativeOne = numberGenerator(-ONE, POSITION, POSITION, SIZES.negativeOne, MARGINS.negativeOne);
+
+    negativeOne.beHappy('start');
+
+    return (
+      <PixiWrapper
+        component={negativeOne.view}
+        isContainer={false}
+        height={SIZES.negativeOne + MARGINS.negativeOne + MARGINS.negativeOne}
+        width={SIZES.negativeOne + MARGINS.negativeOne + MARGINS.negativeOne}
+      />
+    );
+  })
+  .add('sad negative two digits', () => {
+    const negativeTen = numberGenerator(-TEN, POSITION, POSITION, SIZES.negativeTen, MARGINS.negativeTen);
+
+    negativeTen.beSad('start');
+
+    return (
+      <PixiWrapper
+        component={negativeTen.view}
+        isContainer={false}
+        height={SIZES.negativeTen + MARGINS.negativeTen + MARGINS.negativeTen}
+        width={SIZES.negativeTen + MARGINS.negativeTen + MARGINS.negativeTen}
+      />
+    );
+  });