+++ CHANGELOG.md
@@ -1,5 +1,33 @@
 ## HEAD
 
+- 🚀 [2017-05-12] `engine/containers/mazeEngineGenerator`: update to v3.4.0.
+  - 🚀 Update methods to match the new `beSad` number's animation.
+- 🚀 [2017-05-12] `engine/components/numericLineGenerator`: update to v3.2.0.
+  - 🚀 New method: `beSad`. Updates the sad state for each number contained by the `Line`.
+- 🚀 [2017-05-12] `engine/containers/numberGenerator`: update to v3.2.0.
+  - 🚀 Add new method `beSad`: it receives a `state` (`'start' | 'stop'`). Is used when the game has finish uncorrectly.
+- 🚀 [2017-05-12] `engine/helpers/customEases`: update to v1.1.0.
+  - 🚀 Add new ease for the number sad animation.
+- 🐛 [2017-05-11] `engine/containers/numberGenerator`: update to v3.1.1.
+  - 🐛 Fix the stop state for `beHappy` animation. It wasn't resetting its timeline to zero.
+- 🚀 [2017-05-11] `engine/containers/mazeEngineGenerator`: update to v3.3.0.
+  - 🚀 Update methods to match new `beHappy` number's animation.
+- 🚀 [2017-05-11] `engine/components/numericLineGenerator`: update to v3.1.0.
+  - 🐛 Fix `receiveNumberAtPosition`: it wasn't returning the promise.
+  - 🚀 New method: `beHappy`. Updates the happy state for each number contained by the `Line`.
+  - 🚀 `./lineGenerator`: simpler generator function.
+  - 🚀 New stories.
+- 🚀 [2017-05-11] `engine/containers/numberGenerator`: update to v3.1.0.
+  - 🚀 Add new method `beHappy`: it receives a `state` (`'start' | 'stop'`). Is used when the game has finish correctly.
+- 🚀 [2017-05-11] `engine/helpers/customEases`: v1.0.0!
+  - 🚀 Add first custom ease for number happiness animation.
+- 🚀 [2017-05-11] `engine/helpers/randomConfigurations`: update to v1.2.0.
+  - 🚀 Add new export element: `getRandomFloat`. Generate a random float between two numbers.
+- 🚀 [2017-05-10] `engine/helpers/randomConfigurations`: update to v1.1.0.
+  - `randomizeActorsConfig`:
+    - 🚀 New parameter: `difficulty`: is a string that have 3 possible values. It's used to select which range is the one who needs.
+- 🚀 [2017-05-10] `engine/containers/mazeEngineGenerator`: update to v3.2.0.
+  - 🚀 New parameter: `difficulty`: is a string that have 3 possible values and it's used by `randomConfigurations@randomizeActorsConfig`.
 - 🚀 [2017-05-09] `engine/containers/mazeEngineGenerator`: update to v3.1.0.
   - 🚀 Add new validation with new error type (`MazeWrongExitError`) for `numberHasLeftMaze`. This will be thrown if the actual actor's exit is wrong.
   - 🚀 Update `handleResetGame` to match the new `gameMetadata` prop & how actor-exit relation works now.